Implement context-pack service
This commit adds the necessary code to implement the context-pack.service.ts file. This also fixes an import error in app-routing.module.ts. The import statement for CpListComponent was not pointing to the right directory. This commit also changes the field name "$schema" in context-pack.ts and ContextPack.java to "schema" as the dollar sign was causing errors.  Issue #15 Epic #18
Co-authored-by: Richard Lussier <lussi036@morris.umn.edu>
Co-authored-by: Jadyn Sondrol <sondr037@morris.umn.edu>
Co-authored-by: Kyle Day <day00090@morris.umn.edu>
Co-authored-by:  Joshua Eklund <eklun124@umn.edu>