Fix existing tests and remove unnecessary ones
This commit fixes the test AddNullWordList. Originally, the field wordlist was originally left out, but now it has been set to null.  This commit also gets rid of the tests for adding a invalid enabled field and removes the check for that in WordController, as we do not need it. Issue #36 Epic #18
Co-authored-by: Richard Lussier <lussi036@morris.umn.edu>
Co-authored-by: Jadyn Sondrol <sondr037@morris.umn.edu>
Co-authored-by: Kyle Day <day00090@morris.umn.edu>
Co-authored-by:  Joshua Eklund <eklun124@umn.edu>