Add endpoints for context packs for server and remove check
This commit adds the endpoints for the context packs in the server. This commit also removes a check in wordRiverController that checks to see that when adding a new context pack that the wordlist field is not null. We deleted this check and its associated test because you are not required to create a word list when creating an empty context pack. Therefore, there is no need to ensure that the wordlist field is not null when creating an empty context pack.
Co-authored-by: Richard Lussier <lussi036@morris.umn.edu>
Co-authored-by: Jadyn Sondrol <sondr037@morris.umn.edu>
Co-authored-by:  Joshua Eklund <eklun124@umn.edu>