Create tests for cp-info component
This commit creates test in cp-info.component.spec.ts that tests to see if it returns the proper context pack when given an id, and it should return null when given a invalid id. This commit also fixes a naming issue where method names did not match each other.  Issue #17

Co-authored-by: Richard Lussier <lussi036@morris.umn.edu>
Co-authored-by: Jadyn Sondrol <sondr037@morris.umn.edu>
Co-authored-by: Kyle Day <day00090@morris.umn.edu>
Co-authored-by:  Joshua Eklund <eklun124@umn.edu>