Create Cypress test for clicking add context pack button
This commit creates a Cypress test for clicking the add context pack button that is present on the context pack display page. This commit also fixes an issue with the ordering of the HTML that was present in the cp-list.component.html. The button html came before the list display HTML, so the display HTML was covering the button. This commit also seeds the database for packs.json, which gives us basic context packs to preform tests with.  Issue #17

Co-authored-by: Richard Lussier <lussi036@morris.umn.edu>
Co-authored-by: Jadyn Sondrol <sondr037@morris.umn.edu>
Co-authored-by: Kyle Day <day00090@morris.umn.edu>
Co-authored-by:  Joshua Eklund <eklun124@umn.edu>