Test and implement adding empty word lists Server Side
This commits adds a test for adding word lists to context packs. As of right now, the test fails, but this is most likely to due to a syntax error that we don't know how to fix. This commit also implements the addNewWordList() method in WordRiverController and adds an endpoint in the server for adding new word lists.

Co-authored-by: Richard Lussier <lussi036@morris.umn.edu>
Co-authored-by: Jadyn Sondrol <sondr037@morris.umn.edu>
Co-authored-by: Kyle Day <day00090@morris.umn.edu>
Co-authored-by:  Joshua Eklund <eklun124@umn.edu>